some changes

+++ src/main/java/me/plich/cashregistersystem/service/DeviceService.java
@@ -10,14 +10,19 @@ import me.plich.cashregistersystem.repository.LocationRepository;
 import me.plich.cashregistersystem.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
+import javax.persistence.EntityNotFoundException;
 import java.time.LocalDate;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class DeviceService {
@@ -41,12 +46,14 @@ public class DeviceService {
     }
 
     public void deleteDevice(@PathVariable Long id) {
-        Device device = deviceRepository.getOne(id);
-        if (device.getUser().getId() == userService.currentLoggedUserId()) {
+        Optional<Device> deviceOpt = deviceRepository.findById(id);
+        if(deviceOpt.isPresent()) {
+            Device device = deviceOpt.get();
+            if (device.getUser().getId() == userService.currentLoggedUserId()) {
             deviceRepository.deleteById(id);
+            }
         }
     }
-
     public Device getDevice(@PathVariable Long id) {
         Device device = deviceRepository.getOne(id);
         if(device.getUser().getId()==userService.currentLoggedUserId()) {
@@ -54,6 +61,8 @@ public class DeviceService {
         }
         return null;
     }
+
+
 
     public List<Device> getAllUserDevices() {
         List<Device> devices = deviceRepository.findByUser_Id(userService.currentLoggedUserId());
@@ -71,58 +80,61 @@ public class DeviceService {
     }
 
     public void updateDevice(@PathVariable Long id, @RequestBody Device device) {
-        Device deviceToUpdate = deviceRepository.getOne(id);
-        if(deviceToUpdate.getUser().getId()==userService.currentLoggedUserId()) {
-            if(device.getUniqueNumber() != null) {
-                deviceToUpdate.setUniqueNumber(device.getUniqueNumber());
-            }
-            if(device.getSerialNumber() != null) {
-                deviceToUpdate.setSerialNumber(device.getSerialNumber());
-            }
-            if(device.getProducer() != null) {
-                deviceToUpdate.setProducer(device.getProducer());
-            }
-            if(device.getModel() != null) {
-                deviceToUpdate.setModel(device.getModel());
-            }
-            if(device.getDateOfFiscalization() != null) {
-                deviceToUpdate.setDateOfFiscalization(device.getDateOfFiscalization());
-            }
-            if(device.getDateOfDeRegistration() != null) {
-                deviceToUpdate.setDateOfDeRegistration(device.getDateOfDeRegistration());
-            }
-            if(device.getEvidenceNumber() != null) {
-                deviceToUpdate.setEvidenceNumber(device.getEvidenceNumber());
-            }
-            if(device.getReviewsFrequency() != null) {
-                deviceToUpdate.setReviewsFrequency(device.getReviewsFrequency());
-            }
-            if(device.getLastReview() != null) {
-                deviceToUpdate.setLastReview(device.getLastReview());
-            }
-            if(device.getPlannedReview() != null) {
-                deviceToUpdate.setPlannedReview(device.getPlannedReview());
-            }
-            if(device.getDailyReports() != null) {
-                deviceToUpdate.setDailyReports(device.getDailyReports());
-            }
-            if(device.getActive() != null) {
-                deviceToUpdate.setActive(device.getActive());
-            }
-            if(device.getMobile() != null) {
-                deviceToUpdate.setMobile(device.getMobile());
-            }
-            if(device.getReserve() != null) {
-                deviceToUpdate.setReserve(device.getReserve());
-            }
-            if(device.getDescription() != null) {
-                deviceToUpdate.setDescription(device.getDescription());
-            }
-            if(device.getLocation() != null) {
-                deviceToUpdate.setLocation(device.getLocation());
-            }
-            deviceRepository.save(deviceToUpdate);
+        Optional<Device> deviceToUpdateOpt = deviceRepository.findById(id);
+        if (deviceToUpdateOpt.isPresent()) {
+            Device deviceToUpdate = deviceToUpdateOpt.get();
 
+            if (deviceToUpdate.getUser().getId() == userService.currentLoggedUserId()) {
+                if (device.getUniqueNumber() != null) {
+                    deviceToUpdate.setUniqueNumber(device.getUniqueNumber());
+                }
+                if (device.getSerialNumber() != null) {
+                    deviceToUpdate.setSerialNumber(device.getSerialNumber());
+                }
+                if (device.getProducer() != null) {
+                    deviceToUpdate.setProducer(device.getProducer());
+                }
+                if (device.getModel() != null) {
+                    deviceToUpdate.setModel(device.getModel());
+                }
+                if (device.getDateOfFiscalization() != null) {
+                    deviceToUpdate.setDateOfFiscalization(device.getDateOfFiscalization());
+                }
+                if (device.getDateOfDeRegistration() != null) {
+                    deviceToUpdate.setDateOfDeRegistration(device.getDateOfDeRegistration());
+                }
+                if (device.getEvidenceNumber() != null) {
+                    deviceToUpdate.setEvidenceNumber(device.getEvidenceNumber());
+                }
+                if (device.getReviewsFrequency() != null) {
+                    deviceToUpdate.setReviewsFrequency(device.getReviewsFrequency());
+                }
+                if (device.getLastReview() != null) {
+                    deviceToUpdate.setLastReview(device.getLastReview());
+                }
+                if (device.getPlannedReview() != null) {
+                    deviceToUpdate.setPlannedReview(device.getPlannedReview());
+                }
+                if (device.getDailyReports() != null) {
+                    deviceToUpdate.setDailyReports(device.getDailyReports());
+                }
+                if (device.getActive() != null) {
+                    deviceToUpdate.setActive(device.getActive());
+                }
+                if (device.getMobile() != null) {
+                    deviceToUpdate.setMobile(device.getMobile());
+                }
+                if (device.getReserve() != null) {
+                    deviceToUpdate.setReserve(device.getReserve());
+                }
+                if (device.getDescription() != null) {
+                    deviceToUpdate.setDescription(device.getDescription());
+                }
+                if (device.getLocation() != null) {
+                    deviceToUpdate.setLocation(device.getLocation());
+                }
+                deviceRepository.save(deviceToUpdate);
+            }
         }
     }
     public List<Device> findAllByRsql(@RequestParam(value = "search") String search) {
